Store checkpointed tree
I previously thought that editing the tree would only shuffle around
some byteranges. If I kept a copy of those old byte ranges I'd be good.

I now realize edits can remove nodes from the tree. That will result in
problems when we diff trees later. So I'm going to store a clone of the
old tree. This shouldn't be too much of an overhead, since we only clone
when creating a new checkpoint.