Diff against unedited tree
The edit operation on a tree can remove nodes. When diffing we'd like to
keep those nodes in consideration, or otherwise the diff algorithm won't
be able to figure out they were removed.

This means we need to remove the has_changes() check, because it's the
edit operation that marks a node as changed. That likely has performance
implications, and I should consider adding change detection back in a
different way later.