Extract editor listening into a separate thread
My main thread handles change events. Currently the only change events
are the ones coming from the editor, but I'd like the compilation thread
to be able to issue them as well. For this reason I'm creating a more
formal communication channel between the editor listener and the event
handler, to allow that channel to be used by the compilation thread as
well.