Perform source edits on the main thread
There's some work elm-pair has to keep up with, and some work that's
'optional'. The non-optional work is to update the latest code snapshot
with editor events. If we skip editor events elm-pair's local copy of
the source tree will get corrupted.

The optional work is running diffs and acting on them. Doing that kind
of analysis is the point of elm-pair, but if the process is getting
overwhelmed it's okay to do less of it.

This commits moves the non-optional tree upating work to the main
thread, and makes the optional analysis work behave more like the
compilation thread: we run the analysis thread in a loop, doing a single
analysis at a time based on the latest data.

The nice thing about this approach is that if the programmer's machine
allows it we'll get fast analysis in while the programmer is typing, but
if the programmer is typing too fast and/or their machine is not fast
enough to get analysises in the moment, they _will_ happen when the
programmer slows down.