Abstract Editor behind traits
This is a premature change in the sense that I'll probably not get
around to doing another editor integration for a while. The reason for
doing this now is that I want to leverage some of this abstraction to
allow tests to use it. Tests are currently still 'covering' some code
that has been moved (well, copied) into the `neovim` module. If I can
adjust those tests to use the editor API instead, the code they're still
'covering' can be deleted.