Add initial version of knowledge-base
As I'm moving towards building the support for working with multiple
files and buffers I'm going to need to calculate and store information
about these files, and invalidate that information in certain cases in
response to user interaction. This is a first draft of a datastructure
for storing this kind of information.