editor listener thread tracks multiple buffers
This required changes to the editor integration API. The existing API
used separate callbacks for loading and saving source code. To store
multiple buffers both these callbacks needed mutable access to the
editor listener thread state, which is not allowed because Rust can't
tell when those callbacks will get called.

Because on every editor event we read from the first callback then write
to the other, it should be possible to pass the callback a function.
Then:
1. The editor listener loop gets triggered by callback when an editor
   event occurs. It is passed a function to call with the current
   contents of the buffer the event is for.
2. The editor listener loop loads the current code and passed it to the
   function.
3. The editor integration gets the current code and overwrites (parts
   of) it with new content. It returns the new code back to the edit
   listener event.
4. The editor listener event saves the new code.

It was tricky putting this in practice. Rust wanted a concrete type for
the inner function, the one the editor integration provides the editor
listener via callback. To do this the editor integration thread doesn't
provide a function directly, but rather a value of a new associated type
`EditorEvent` on the Editor trait. `EditorEvent` has a single method the
editor listener thread needs to call.

The `EditorEvent` trait needs access to some properties in the editor
itself. For example: the reader it should read source code from. Ideally
the `EditorEvent` would hold a reference to this data, since it's very
short-lived. However, this would require `EditorEvent` to take a
lifetime parameter. Rust currently requires parameters on associated
types to be put on the trait too, but this would mean each `EditorEvent`
needs to live as long as the `Editor`, which isn't the case. Support for
Generic Associated Types would probably enable this, but it's not there
yet:
https://github.com/rust-lang/rfcs/blob/master/text/1598-generic_associated_types.md

If `EditorEvent` can't reference the reader then it needs to own it. The
Editor will need it back after the EditorEvent is discarded, to parse
the next `EditorEvent`. To make this work I let some editor methods take
consume the editor and return a new one (conceptually, very little
copying actually takes place).