Park cornercase Dict.Dict for now
If a module name and the name of a type exposed from it are the same,
then diffing can lead to ambiguous results. For instance, if we see a
`Dict` being changed to `Dict.Dict`, that means a qualifier was added.
`elm-pair` currently smises `.Dict` was added, which makes no sense.

The best fix for this, I think, is to change the tree-sitter grammar to
ensure that qualifications and typenames use different node types. I
want to do that, but not right this moment. This commit sets aside the
test case that (accidentally) discovered the bug and creates a new,
simpler test case to focus on now.