Fix add_module_qualifier_to_type_with_same_name
This required a change in the tree-sitter-elm library, to help
distringuish module names from Elm names. Without this, when changing
`Dict` to `Dict.Dict`, the tree-diffing logic is not able to understand
whether we added `Dict.` before the existing `Dict`, or `.Dict` after
the existing `Dict`. When both `Dict`s have different node kinds, then
the diffing logic is able to see which is the new one.