Combine Errors back into a single type
Having tried both a single monolithic error type and a split up one, I
find the monolithic one easier to work with. It saves a lot of
conversion work.

Philosphically it's not that weird either. Each of the error
constructors represents a cause that currently brings down the entire
elm-pair process. In that sense these are all top-level errors. If that
doesn't make sense for certain errors then the problem is now more
visible.

The immediate need here is sharing logic (and so Error return values)
beteween the analysis and compilation thread. If they're already sharing
the same error type then there's no work here.