Parse i.dat file
This file contains information about exposed values of Elm dependencies.
This information is needed to, for instance, know that when `Result(..)`
is removed from an exposed-list on an import then the constructors `Ok`
and `Err` will go out of scope.

The information we read isn't hooked into the application logic yet and
as a result some tests are failing. Fixing this is the next order of
business.