Push more info on exposed values into types
This is preparing for dealing with `exposing (..)`. Directly
implementation proved tricky because in various places we assumed
exposed values had some generic properties like a 'name', that don't
hold for `(..)`. Pushing more information in the types and forcing
existing code to deal with it will let the compiler guide us to the
breakages when we add a new kind of exposed value.