refactor code does not mutate KnowledgeBase
Previously we'd insert data into the KnowledgeBase whenever it's first
requested. It's a nice idea, but it requires any code querying the
knowledgebase to hold an exclusive reference to it, given any query
might turn into an update. This was causing increasing amounts of
issues with the borrow checker.

This commit makes it so we insert information into the KnowledgeBase
proactively, when we open a new buffer for instance. That way the
refactor code only needs to read from it.