Store all elm-pair files in same directory
This directory should end up in the user's home datadirectory, but
that's a different location per platform. Once the plugin learns the
socket location from invoking elm-pair as a CLI, then elm-pair will be
the only source of truth for directory locations, and the move of the
elm-pair directory will then be easy to make.