Update last_compiling_version after refactor
This fixes a problem where elm-pair cannot keep up, for example when
changing the as-clause of an import. What would happen is that the first
change to the as-clause would register, then subsequent refactors would
fail because additional changes would still compare against the last
compiling version, which hadn't changed.