Give refactor code versions their own revisions
This is intended to help with debugging. If we don't do this the
original code state from the editor and the changed one after the
refactor will have the same revision numbrs, which can create
ambiguities.

We want to ensure that if we receive a new change from the editor, it
will have a unique revision as well, and a higher one than the revision
of our previous refactor. By using even numbers for editor changes and
odd ones for refactors we can guarantee this.