Neovim plugin starts elm-pair daemon
This makes it so it's no longer necessary to start an elm-pair process
manually before opening the editor. The plugin will now start elm-pair
directly. Elm-pair will then point the editor to the socket it can
connect to.

Now that the socket path is no longer hardcoded in the plugin, we can
let the elm-pair process figure out the proper place to store
application helper files in the user's home and put its files there,
rather than under /tmp. This commit doesn't make that change though.