Add elm binary to elm-pair closure in Nix install
This removes the potential error of elm-pair not running because the elm
binary cannot be found, at least when using Nix. When we get to
implementing other installation methods we'll have to figure out what to
do where.

This does couple elm-pair more tightly to a particular elm version, but
it was coupled tightly already (because of the elm parser we're using).
Figuring out what to do when multiple Elm versions are in play is left
for future consideration. Elm doesn't update all that often, so I don't
consider it an urgent question.