Only record type alias conflicts with constructor
This Elm code compiles:

    type alias Foo = Int

    type Bar = Foo

But this code does not:

    type alias Foo = { a : Int }

    type Bar = Foo

The latter example fails because `Foo` can both be used as a constructor
for the `{ a : Int }` record as well as a constructor of a the `Bar`
type. In the first example there's no such ambiguity.

Before this code always assumed a conflict between a type alias name and
a constructor name. Now we only assume such as conflict when the type
alias is for a record.