Test conflicting module exports when exposing all
When we expose a value from one import with the same name as a value
already exposed from another import, then elm-pair will drom the value
from the exposing list from the other import and qualify it's usage. The
above statement is true with one exception: if the import is exposing
everything `(..)`, then we do nothing. The result is non-compiling code.

This test (currently failing) demonstrates the desired behavior to
implement: The `exposing (..)` should be replaced with an exposing list
containing all values we're using from the module, with the exception of
the value we want to qualify.