Reintroduce PathBuf types
To let PathBuf's flow through differential-dataflow it needs to
implement the `Abomonation` trait. `abomonation` is a fast
serialization/deserialization library in the timely dataflow ecosystem.

I've made a local fork of the library and added the trait implementation
there. That seems to work. I'm going to try to contribute the change
back upstream.