Switch to experimental version of naersk
Without this the Nix build is currently failing, because the stable
naersk has a bug related to using git dependencies in crate overrides,
see:

https://github.com/nix-community/naersk/issues/152

Fixing to the naersk in this PR fixed things:

https://github.com/nix-community/naersk/pull/167

This should be a short-lived changed. I hope to either get my PR for the
abomonation library merged which would make the crate override
unnecessary. Should it not get merged I'll likely abandon my abomonation
fork for some alternative solution.