Implement incremental fileparsing
This creates the basis of incremental fileparsing. When a single file
changes, only that file will be reparsed.

This reverts to using the currently stable version of the `notify`
crate. Motivation for switching to the unstable branch was that it
allowed passing a callback function to the file watcher, rather than a
channel. If that callback function is to insert file events directly
into the dataflow, then it will need a reference to the dataflow inputs.
Those inputs do not implement `Send` though, and this is needed because
the `notify` gem runs the file watcher on a separate thread.

The code still needs some cleaning up. There's a bunch of TODO's,
`unwrap()`s, and unnecessary cloning going on. I also want to add some
tests for the dataflow logic.

Paired with @jeffawang.