Try compiling without musl for now
I like the idea of having static binaries. The current obstacle to
making progress is tree-sitter-elm having a C++ component (the custom
scanner). There might be a way to get that to work with Musl, and there
s also the option of writing that scanner in C. Either way, I don't want
to spend the time on that now.