Find Elm compiler once and reuse the result
Previously the compiler path to used was determined at compile time. The
plan is to change this to find a compiler next to the elm-pair binary,
put there by the editor plugin. To run this logic once rather than once
per compilation action we extract it to the root of the application.
This has some added advantages:

- If we can't find an Elm compiler we can exit early.
- If we want to add locking behavior at one point to ensure only one
  compilation is underway at once, that's now easier to set up.