Turn ProjectInfo type into alias
This is a step towards disolving the type entirely. The code using the
ProjectInfo type is never interested in the entire thing. They have
more specific questions that often can be answered by a single module.
If we can encode the questions into the dataflow computation, then we
don't need the container type anymore.