Don't construct project HashMap
We used to construct a giant HashMap containing all the modules in all
projects being tracked by elm-pair. It needed to be rebuilt any time a
file in any project changed. No consumer of this information used the
whole thing, so it was a bit wasteful to construct.

We now only construct information about individual projects. We let
dataflow store the result (it's doing so already anyway), and when we
need information about a project we query it out of dataflow's internal
aggregate structure.

The goal is to break this down further by letting dataflow calculate the
information we need, rather than some in-between datastructure
representing all there is to know about a project, from which we then
pull the information we need.