Pass collections into dataflow graph
The existing dataflow graph, which is concerned with maintaining an
overview of which files are in which project, now takes dataflow
collections and returns aggregates and probes. This will make it
possible for tests to work on this bit in isolation. As I'm adding more
stuff to dataflow, I won't need to throw it all in the same function.