Pull tests loose from dataflow evaluation logic
I like the current dataflow logic as a unit and want to keep the
existing tests for that unit. At the same time I want to pull more logic
into dataflow. This commit unties the dataflow tests from the
boilerplate used to run the dataflow computation. That way the
boilerplate can expand to include more dataflow logic without needing to
alter the tests to cover this logic as well; I'd like to write separate
tests for that.

The dataflow advancement logic has been a source of some bugs in the
past. I'd like to share that logic between tests and real code, so tests
cover it. It's been generalized to make this possible.