Scope checks only test node begins in scope range
The logic that tested whether a certain parameter fell within a
particular scope tested that both the node start and end byte fell
within the range that defined the scope. That makes sense, but we're
doing more work than is necessary. Given that the ranges of nodes and
scopes both are extracted from parse tree nodes, we know these ranges
will never overlap, like so:

    [---- range a ----]
            [---- range b ----]

Given such overlap is impossible, checking whether the start byte of the
supposed contained range lies within the supposed containing range is
enough.