Avoid allocating unnecessary Vec
This avoids allocating an intermediary Vec to contain the data read from
the idat_parsing logic. This might help address the following issue:

https://github.com/jwoudenberg/elm-pair/issues/6

I'm not entirely convinced this will be enough. The data that was in the
Vec causing the crash above will now be written into a different Vec as
part of the differential-dataflow `reduce` operation. It's not unlikely
the error will return with a different stack trace. That would also
provide useful info though, showing us we're on the right path. Also,
dropping an intermediary allocation is in itself a nice improvement.