Only trigger rename when we change name definition
Previously we could either change a name at the place where it is
defined or a place where it is being used, and either would cause
Elm-pair to trigger a rename. It turns out this often leads to Elm-pair
misunderstanding programmer intent. Instead of renaming, the programmer
might be trying to call a different function, use a different
constructor, or a different type.

When we change the name at the definition site we can be more sure a
rename is what the programmer intends, but the downside is that the
programmer will have to navigate to the definition site first.

Adresses: https://github.com/jwoudenberg/elm-pair/issues/8