Avoid undo loops
If the undo of an Elm refactor is itself a change that results in an
Elm-refactor, than hitting undo can get the programmer into a loop: Each
undo removes the previous change, but then Elm-pair adds a new one. This
commit adds protection against this.

The idea is that we store the version of the code prior to the previous
refactor, and if a change directly on top of this refactor takes us back
to it we know it's an undo.