Improve undo-loop detection
The previous undo logic provided a lot of false positives: it thought
changes were undo's that weren't. Each time an undo was detected it
would forget the last compiling version. If the user was typing fast
enough they might alread have introduced new uncompiling changes, and
with Elm-pair not having a last compiling version to compare against it
would do nothing in response to these changes. The programmer would
notice this as Elm-pair dropping the thread, as in this issue:

https://github.com/jwoudenberg/elm-pair/issues/12

The new logic doesn't try to detect undo's. Instead, it tracks whether a
refactor Elm-pair is about to make is one of the previous two refactors
it made, and if so it aborts. It's simpler and not as hard to fool.