Expand type-check scanning of expression types
The type checker isn't smart enough yet to understand that these are the
same:

   x `SameAs` ArgTo(y)
   x `ArgTo` y

And similar with `ResultOf`:

   x `SameAs` ResultOf(y)
   x `ResultOf` y

For this reason it's not able to formulate any rule matching:

  * `ResultOf` List.length

This means `List.length` is not discovered as a function type, and so
the link between it with the `(++)` operator not found either. Once this
is fixed the type of `(++)` should be inferred.