Support multiple variable scopes
Up to this point the type checking code assumed all variable names lived
in the same scope. This commit adds a test where two variable with the
same names have different types in different scopes, and the code to
make type checking support this scenario.

The approach taken is to pass a stack of current scopes along our
tree-sitter scanning logic. When we encounter a variable definition we
attach it to the innermost scope. When we encounter a variable usage we
add attach it to hypothetical definitions in every scope on the stack.
Because Elm does not allow variable shadowing only one of those
definitions will actually exist during type checking.