Communicate an editor-kind id to Elm-pair
Elm-pair can use this to identify the kind of editor that is connecting
to it, and run appropriate server logic.

Note that having a single protocol for communication between editor and
Elm-pair is a not a design goal of Elm-pair. Instead the goal is to
implement some per-editor custom logic into the Elm-pair codebase. The
reasons for this approach are:

- The assumption better integrations are possible by playing to each
  editor's strenghts. For example, for Neovim msgpack-rpc is the best
  supported and most performant way to integrate so that's what we use.
- The assumption that this approach requires less code and complexity in
  the editor plugins themselves (as opposed to the elm-pair codebase).
  Suppose we'd design one binary protocol for communication that all
  editors would need to use, this would require the use of custom code
  and/or dependencies in editors that do not support the chosen
  protocol.