Bundle vscode extension using Nix
This is useful for Nix users that wish to plug Elm-pair into a
Nix-managed vscode configuration, but the bundle Nix builds will also
used by me as the first step towards publishing the vscode plugin on the
vscode marketplace.

I might eventually want to do the same for the neovim plugin, but
instead of calling some publish command push it into a dedicated github
repository (given that's more or less the stanard for Vim plugin
managers).