Add test for vscode extension
This test covers the internal logic of the extension, but not
interaction with the visual studio code API. All else being equal it'd
be nice to cover that too, but I do believe it would make the complexity
of the tests go up by a lot because they'd need to become integration
tests. Given the interaction with vs-code API changes infrequently, and
can be manually tested performing a single refactor it isn't worth it
for me to look into this now. I like the no-dependency setup of the
current tests.