don't refactor in response to vscode undo
Vscode is able to tell for each change whether it was the result of an
undo/redo operation or not. We can reuse this to reliably prevent
Elm-pair from refactoring in response to undo/redo changes. Neovim
unfortunately does not provide us this information, so we have to fall
back to less reliable undo-detection for it.