Switch to Zola for site generation
This is preparation for future changes. I'd like to add a bunch of
additional pages soon, and some templating seems nice to be able to
share a basic layout and styling between the various pages.

This will rename the site feed from /feed.xml to /atom.xml. I don't
particularly like this, but haven't found an easy way to keep the old
url and don't want to invest more time in it.

I'm going with Zola being it's a single-binary dependency and has the
features I need, including an RSS feed.