Fix renaming issue
When renaming a variable to a name that's already in use in a different
scope, before this commit Elm-pair would rename that other variable even
though it needn't (because it's in a different scope). This commit fixes
the issue, and covers it in a test.

This fixes https://github.com/jwoudenberg/elm-pair/issues/13.