Make static site buildable through nix
Not super useful, but I'm on a train without an internet connection and
I wanted to try this out.

It'd be nice at some point if the flake.nix could be used to build all
components of this projects. The derivation containing the static site
artifacts could be pushed to something like github pages, the elm-pair
binaries to a release, etc. Current obstacle is that the elm-pair
binaries reference dynamacially linked libraries linked to
/nix/store-addresses, so those binaries cannot be distributed directly.
Something I'd like to look into at some point, for instance by building
statically linked artifacts.