When no project buffer can be found, don't crash
To be able to compile Elm code we need to know the project root of the
buffer we're trying to compile. Currently elm-pair crashes when this
directory cannot be found. That spoils the party for other buffers that
_do_ exist.

Fixes https://github.com/jwoudenberg/elm-pair/issues/14

There's other reasons a project root might not be found though. The
programmer might have created a small napkin Elm file to try something
out for instance. It's reasonable for Elm-pair to log an error in that
scenario, but crashing would not be helpful.