Add sales page
This page is laying groundwork for the sale of commercial Elm-pair
licenses. I need to put this live so I can get myself verified on
paddle.com (the payment solution I'd like to use). None of this is live
yet, hence the disclaimer at the top of the page and no links leading
to the page yet.