Add elm-pair --credits command
This prints license information for dependencies, ensuring those
licenses are included in binary distributions of elm-pair.

This includes all (indirect) dependencies under the MIT or ISC license.
For dependencies dual-licensed under MIT and Apache-2.0 I choose to
distribute under the MIT license. Dependencies under BSL-1.0 and the
Unlicense do not require attribution to be present in binary
distributions.