Move editor integrations into subdirectory
The side-purpose of this change is to intentionally break folks who set
up their vim plugin to track this repository (sorry!). Because this new
version changes the licensing I want to make sure folks aware so they
don't break the new license by accident when they upgrade the plugin.
This change will force folks to return to the repo to look up the new
plugin location, allowing them to learn about the new license in the
process.