Access file once per file-based test
In file based tests a test file contains an input and an expected output
section. Before this commit every individual test was responsible for
reading the input portion of the test out of the test file. Now the test
helper does this. This removes an extra read from the file (though this
does not seem to make an impact on test performance), and simplifies the
implementations of individual tests.