VS-code reports initial opened file to elm-pair
The first Elm file opened in Visual Studio Code will activate the
extension, which will register handlers to listen for additional Elm
files opened later. That first file won't be registered though until we
make a change in it. If that change is something Elm-pair could act
upon, that action will not take place.