WIP: Open other source files in the editor
This makes some progress towards Elm-pair being able to open additional
source files in the programmer's editor.

Up to this point Elm-pair only ever edited whatever file the programmer
currently had open. We want to overcome this limitation to allow
Elm-pair to rename names that are exported to other source files. If
some of the source files involved in a rename are not yet open in the
editor, we should open them first. Then, when all Elm-pair has open
buffers for all files it needs to change, it can perform the refactor it
wants to do.

An alternative approach would be for Elm-pair to directly write to the
filesystem changes for files that aren't open in the editor. This would
avoid the need for Elm-pair to open any new files. There's a couple of
reasons I'm not at this time persuing that option:

- If some edits take place in editor buffers and some on the file
  system then in case of an error it's possible for only part of the
  refactor to take place, leaving the programmer in a messy state.
- Saving part of the refactor to the file system, and keeping another
  one in memory in the form of changed editor buffers is arguably
  confusing: Elm-pair would make a refactor and save/persist half of it.
  Pressing undo in the editor would only undo the in-memory part of the
  refactor.
- The way renames currently work, a fast machine might perform a single
  rename for every character typed while performing the rename. This
  might result in performance problems if each of those little renames
  results in multiple writes to the filesystem. Note: the
  rename-as-you-type effect is not supper important to me, and I'd be
  fine if the rename happens once, when you're done typing. Implementing
  that requires finding a heuristic to detect when the programmer is
  done typing, so more work.