Analysis thread knows all open buffers
Up to this commit the analysis thread only had access to the
last-modified buffer. To perform refactors that modify multiple files
(like renames across files) the analysis thread needs to know the code
for all affected buffers. This commit shares the whole set between the
editor listener and the analysis thread.

As an unintended side-effect this seems to simplify the code a bit as
well. I don't believe there's a performance trade-off here.

As another unintended side-effect, we can remove the last use of the
`mvar` type, as a Mutex has all the functionality we need for sharing
buffer state between the editor listener and analysis threads.