Stop sharing buffer memory between threads
Up to this commit the analysis and editor listener threads shared a
single hashmap containing source code of all active buffers. To avoid
conflicts this hashmap was in a Mutex. This commit makes it so each
thread stores its own source code datastructures. Each time source code
gets edited the editor listener thread will send a clone of the new
source code to the analysis thread over a channel.

This change simplifies the code at the cost of a lot of additional
source code datastructure cloning operations. This should be fine
because the Rope datastructure containing source code is an immutable
datastructure that is very cheap to clone, according to its
documentation.

The direct purpose of this change is supporting cross-file renames. This
will require the refactor code to have access to source code of not just
the last changed buffer, but potentially other buffers as well. The old
set-up where we would lock the Mutex of just the active buffer briefly
just before a Mutex, just long enough to make a clone of that single
buffer, no longer works. We would need to either clone _all_ the buffers
at that moment to prepare for us potentially requiring access later, or
pass the mutex containing all buffers down in to the refactor logic and
lock it again later if we need data about additional buffers.