Open dependent modules when renaming
This makes it so that when we change the name in a module that is
imported by other modules, we open those other modules instead of doing
the refactor. This is a step towards renaming across files.

Apart from the rename-across-files logic itself, the following bits are
missing:
- Don't open buffers that area already open
- Don't open buffers for which the command to open has just been sent
- Don't open buffers if we're not exporting the renamed value
- Don't open buffers that do not import the renamed value

An additional complexity is that the logic above isn't at the moment
coverable by existing tests, so we'll have to address that too.