First vertical integration of cross-file renaming
This is an initial implementation of renaming across files, albeit still
limiting and buggy. It only renames unqualified values (so no
Foo.old_name), only in Neovim, and on top of that renames in a bunch of
cases where it shouldn't. We'll have to work on all those cases in
upcoming commits.