Run simulation tests against analysis thread API
The makes the tests cover more code, at the cost of making them slower.
One test is now failing because it uncovered a legit bug: The logic for
detecting invalid Elm code prior to comitting to a refactor is not
working. This bug was previously uncovered in #18, it's nice to have a
test for it now. The test will be fixed in a future commit.

The immediate reason for making this change is to allow writing tests
that check cross-file changes. Such tests require the
dependency-tracking logic to be in scope. An alternative approach not
taken would be to add a mock implementation for the dependency tracking
logic.

I've wanted to test against the analysis thread API before, but it used
to be harder when that API included some shared state in addition to the
message bus. A couple commits back we went to just having the message
bus, making this change much easier.