Allow passing non-refactor messages to vscode
Up to this point we were assuming all messages passed from the elm-pair
process to the visual studio code extension where refactor payloads.
We're going to need to support 'open these files'-style commands as
well. This commit changes the communication protocol to include a
'commandId' for distinguishing these different types of commands.