Initial version of licensing server
From speaking to folks I'm getting some signals that the current
Elm-pair licensing situation is unclear. Part of it has to do with
Elm-pair itself having no idea whether it's running a commercial license
or not. It's entirely up to folks using it to educate themselves about
whether they need a commercial license, and then to buy one of so.

I'd like for Elm-pair to gently nudge users towards the commercial
license if they don't have one, but this requires Elm-pair to know
whether it is licensed. My current plan is to introduce a license file
for this that Elm-pair will check once on start-up and validate without
making any HTTP requests. This requires me to start generating license
keys.

Paddle.com, my payment provider, is set-up for this, and can call a
webhook to generate a license after a succesfull order. This commit
implements the first part of a licensing server that will respond to
such a request.