Show license information when starting elm-pair
This displays a page with information about Elm-pair licensing when
starting Elm-pair. I'm hoping this will nudge a few folks that might be
using Elm-pair for commercial work to get a license.

The work here is incomplete. License activation, as described in the
information text, is not implemented yet. This will follow in a
subsequent commit.