Stop using abomonation for timely serialization
It looks like it's not super well maintained at the moment, and upstream
seems to attempt to move away from it towards 'serde'. Serde has some
performance trade-offs, especially when it comes to types that need to
be cloned, like strings. I got some thoughts on cutting down on the
amount of strings flowing through the dataflow computation, we'll see if
those are needed.