Update credits.txt
Some additions are due to the the recent cargo update, but most are
related to support for platforms (Windows, Fuchsia, Zircon) that
Elm-pair itself currently does not support. Strictly these aren't
dependencies because we never compile for these targets, but it's easier
to add them to credits.txt then to filter them out in the
check-credits.sh script.